+++ src/Model/Top.php
@@ -1,0 +1,11 @@
+<?php
+
+namespace Jikan\Model;
+
+
+class Top extends Model
+{
+
+	public $top = [];
+
+}